+++ packages/client/index.html
@@ -8,6 +8,6 @@
   </head>  
   <body>
     <div id="root"><!--ssr-outlet--></div>
-    <script type="module" src="./src/main.tsx"></script>
+    <script type="module" src="./src/entry-client.tsx"></script>
   </body>
 </html>
