+++ packages/client/ssr.config.ts
@@ -17,7 +17,7 @@ export default defineConfig({
   },
   build: {
     lib: {
-      entry: path.resolve(__dirname, 'ssr.tsx'),
+      entry: path.resolve(__dirname, 'src/entry-server.tsx'),
       name: 'Client',
       formats: ['cjs']
     },
